+++ geopandas/io/tests/test_infer_schema.py
@@ -10,6 +10,7 @@ from shapely.geometry import (
 )
 
 import pandas as pd
+import pytest
 import numpy as np
 from geopandas import GeoDataFrame
 from geopandas.io.file import infer_schema
@@ -278,12 +279,26 @@ def test_infer_schema_null_geometry_all():
     assert infer_schema(df) == {"geometry": "Unknown", "properties": OrderedDict()}
 
 
-def test_infer_schema_int64():
-    int64col = pd.array([1, np.nan], dtype=pd.Int64Dtype())
+@pytest.mark.parametrize(
+    "array_data,dtype", [([1, 2**31 - 1], np.int32), ([1, np.nan], pd.Int32Dtype())]
+)
+def test_infer_schema_int32(array_data, dtype):
+    int32col = pd.array(data=array_data, dtype=dtype)
     df = GeoDataFrame(geometry=[city_hall_entrance, city_hall_balcony])
-    df["int64"] = int64col
+    df["int32_column"] = int32col
 
     assert infer_schema(df) == {
         "geometry": "Point",
-        "properties": OrderedDict([("int64", "int")]),
+        "properties": OrderedDict([("int32_column", "int32")]),
+    }
+
+
+def test_infer_schema_int64():
+    int64col = pd.array([1, np.nan], dtype=pd.Int64Dtype())
+    df = GeoDataFrame(geometry=[city_hall_entrance, city_hall_balcony])
+    df["int64_column"] = int64col
+
+    assert infer_schema(df) == {
+        "geometry": "Point",
+        "properties": OrderedDict([("int64_column", "int")]),
     }